WIP(integer-i64): widen codegen emissions + stdlib const integer
Stage 2 of the i64 migration on int_migrate.  Build now passes
(both library and tests compile).  Runtime SIGSEGVs on a minimal
`println("hello")` — the drift is documented below and needs
follow-up.

Codegen (src/state/codegen.rs):
- OpCall / OpCoroutineCreate back-patch offset: pos+7 → pos+11
  (d_nr widened from 4 → 8 bytes).
- OpCall emission: `op as i32` → `op as i64`,
  `code_position as i32` → `as i64`.
- OpConstStoreText: `rec as i32` / `0i32` → `i64::from(rec)` / `0i64`.
- add_const Type::Integer arm: `code_add(*nr)` (i32, 4 bytes) →
  `code_add(i64::from(*nr))` (8 bytes).  Catches every operator
  that takes a `const integer` argument via the attribute walk.

Stdlib (default/*.loft):
- OpCall / OpConstLongText / OpCoroutineCreate signatures:
  `const i32` → `const integer` (now 8 bytes consistently).

Dump/debug (src/state/debug.rs):
- dump_attribute Type::Integer reads i64 (was i32).
- fn_name (both callers) reads addr as i64 (was i32).

Downstream Rust fns widened to accept i64 (matches generator's
view of `integer`-typed params):
- src/ops.rs: `text_character(from: i64)`.
- src/state/text.rs: `string_from_texts(_start: i64, _size: i64)`.
- src/state/mod.rs: `fn_call(to: i64)`.
- src/vector.rs: `get_vector(from: i64)`, `insert_vector(index: i64)`,
  `remove_vector(index: i64)`.
- All call sites updated (parallel.rs, database/search.rs,
  codegen_runtime.rs, state/io.rs).

Tests:
- tests/threading.rs: packed i32 vector elements → `set_i32_raw`.
- tests/data_structures.rs: `Vec<(String, i32)>` → `Vec<(String, i64)>`
  for integer hash values; tree test scaffold → `set_i32_raw`.
- tests/expressions_auto_convert.rs: `2_i32` → `2_i64`.

Fill.rs (generated + hand-patched):
- Regenerated with new rust_type signatures.  Hand-patched 7 sites
  the generator doesn't yet handle (set_ref/set_character/set_text
  use set_u32_raw; get_int null sentinel → i64::MIN; set_byte/
  set_short cast v_val: i64 → i32).

Known runtime failure (needs root-cause fix):
- Minimal `println("hello")` SIGSEGVs in the interpreter.
- Symptom: `op=143 (append_text) at pc=1442, d_nr=346` — the
  bytecode stream is out of sync and dispatch reads random bytes as
  an opcode.
- Likely cause: some remaining place still emits/reads integer
  params with wrong width.  The trace did not pinpoint it.

+++ src/codegen_runtime.rs
@@ -304,7 +304,7 @@ pub fn OpInsertVector(
     index: i32,
     db_tp: i32,
 ) -> DbRef {
-    let new_value = vector::insert_vector(&data, size as u32, index, &mut stores.allocations);
+    let new_value = vector::insert_vector(&data, size as u32, i64::from(index), &mut stores.allocations);
     stores.set_default_value(db_tp as u16, &new_value);
     new_value
 }
@@ -1169,7 +1169,7 @@ pub fn OpRemove<S: IterState>(stores: &mut Stores, state: &mut S, data: DbRef, o
             // plain vector: arg is the element type index
             let elem_size = u32::from(stores.size(arg as u16));
             let n = if reverse { cur + 1 } else { cur - 1 };
-            vector::remove_vector(&data, elem_size, cur, &mut stores.allocations);
+            vector::remove_vector(&data, elem_size, i64::from(cur), &mut stores.allocations);
             state.set_cur(n);
         }
         1 => {
@@ -1217,7 +1217,7 @@ pub fn OpRemove<S: IterState>(stores: &mut Stores, state: &mut S, data: DbRef, o
                 return;
             }
             let n = if reverse { cur + 1 } else { cur - 1 };
-            vector::remove_vector(&data, arg as u32, cur, &mut stores.allocations);
+            vector::remove_vector(&data, arg as u32, i64::from(cur), &mut stores.allocations);
             state.set_cur(n);
         }
         _ => {}
